+++ tests/puppeteer/campaigns/functional/BO/catalog/brandsAndSuppliers/brands/03_CRUDBrandAndAddress.js
@@ -9,10 +9,10 @@ const loginCommon = require('@commonTests/loginBO');
 const BOBasePage = require('@pages/BO/BObasePage');
 const LoginPage = require('@pages/BO/login');
 const DashboardPage = require('@pages/BO/dashboard');
-const BrandsPage = require('@pages/BO/brands');
-const AddBrandPage = require('@pages/BO/addBrand');
-const ViewBrandPage = require('@pages/BO/viewBrand');
-const AddBrandAddressPage = require('@pages/BO/addBrandAddress');
+const BrandsPage = require('@pages/BO/catalog/brandsAndSuppliers/brands');
+const AddBrandPage = require('@pages/BO/catalog/brandsAndSuppliers/addBrand');
+const ViewBrandPage = require('@pages/BO/catalog/brandsAndSuppliers/viewBrand');
+const AddBrandAddressPage = require('@pages/BO/catalog/brandsAndSuppliers/addBrandAddress');
 
 let browser;
 let page;
@@ -62,7 +62,7 @@ describe('Create, Update and Delete Brand and Address', async () => {
   // GO to Brands Page
   it('should go to brands page', async function () {
     await this.pageObjects.boBasePage.goToSubMenu(
-      this.pageObjects.boBasePage.productsParentLink,
+      this.pageObjects.boBasePage.catalogParentLink,
       this.pageObjects.boBasePage.brandsAndSuppliersLink,
     );
     await this.pageObjects.boBasePage.closeSfToolBar();
