+++ tests/puppeteer/campaigns/functional/BO/catalog/brandsAndSuppliers/brands/05_BulkActionsBrandAddresses.js
@@ -7,8 +7,8 @@ const loginCommon = require('@commonTests/loginBO');
 const BOBasePage = require('@pages/BO/BObasePage');
 const LoginPage = require('@pages/BO/login');
 const DashboardPage = require('@pages/BO/dashboard');
-const BrandsPage = require('@pages/BO/brands');
-const AddBrandAddressPage = require('@pages/BO/addBrandAddress');
+const BrandsPage = require('@pages/BO/catalog/brandsAndSuppliers/brands');
+const AddBrandAddressPage = require('@pages/BO/catalog/brandsAndSuppliers/addBrandAddress');
 
 let browser;
 let page;
@@ -46,7 +46,7 @@ describe('Create 2 brand Addresses and delete with bulk actions', async () => {
   // GO to Brands Page
   it('should go to brands page', async function () {
     await this.pageObjects.boBasePage.goToSubMenu(
-      this.pageObjects.boBasePage.productsParentLink,
+      this.pageObjects.boBasePage.catalogParentLink,
       this.pageObjects.boBasePage.brandsAndSuppliersLink,
     );
     await this.pageObjects.boBasePage.closeSfToolBar();
