+++ tests/puppeteer/pages/BO/catalog/brandsAndSuppliers/addBrand.js
