+++ tests/puppeteer/pages/BO/catalog/brandsAndSuppliers/brands.js
