+++ tests/puppeteer/pages/BO/catalog/brandsAndSuppliers/viewBrand.js
